Add: Collatable,ActionableCollatable,Risk,Action
Here is where I am getting started on the class and interface files
for the collations. I'm just going off of the db schema prototype.
This is a little gross, but will do for now.

After getting the objects finished (next commit), the next task is to
verify db connection works for the objects. Hopefully, we will be able
to recast them after getting an array of generic collatables from arango.